services/sftp-watcher/requirements.txt
kafka-python==2.0.2
paramiko==3.4.0